refactor(modules): move configure hook from ViewModule to ElectronLifecycleModule
The configure hook (noAsar, electron flags, data path redirection) is
Electron runtime configuration unrelated to view management. Move it to
ElectronLifecycleModule which already owns await-ready and quit.
